re-factored code for more modularity
in order for me to maintain and add additional features, going to a modular approach seemed better. (Time will tell)
also experimenting with not using classes and more static factory concepts.
The whole code was reflected but this did bring about some major issues with overlays working in the paint, difficult to trace, I suspect the rework of the U++ code in this area would be a good idea ( very painful)